+++ src/main/java/fuzs/pickupnotifier/mixin/client/accessor/AbstractArrowAccessor.java
@@ -10,5 +10,4 @@ public interface AbstractArrowAccessor {
 
     @Invoker
     ItemStack callGetPickupItem();
-
 }