+++ src/main/java/fuzs/pickupnotifier/network/message/S2CTakeItemMessage.java
@@ -1,11 +1,11 @@
 package fuzs.pickupnotifier.network.message;
 
 import fuzs.pickupnotifier.client.handler.AddEntriesHandler;
+import fuzs.puzzleslib.network.message.Message;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraftforge.fmllegacy.network.NetworkEvent;
+import net.minecraft.world.entity.player.Player;
 
-public class S2CTakeItemMessage implements IMessage {
-
+public class S2CTakeItemMessage implements Message {
     private int itemId;
     private int amount;
 
@@ -14,29 +14,32 @@ public class S2CTakeItemMessage implements IMessage {
     }
 
     public S2CTakeItemMessage(int itemId, int amount) {
-
         this.itemId = itemId;
         this.amount = amount;
     }
 
     @Override
     public void write(FriendlyByteBuf buf) {
-
         buf.writeVarInt(this.itemId);
         buf.writeVarInt(this.amount);
     }
 
     @Override
     public void read(FriendlyByteBuf buf) {
-
         this.itemId = buf.readVarInt();
         this.amount = buf.readVarInt();
     }
 
     @Override
-    public void handle(NetworkEvent.Context ctx) {
-
-        ctx.enqueueWork(() -> AddEntriesHandler.addPickUpEntry(this.itemId, this.amount));
+    public TakeItemHandler makeHandler() {
+        return new TakeItemHandler();
     }
 
+    private static class TakeItemHandler extends PacketHandler<S2CTakeItemMessage> {
+
+        @Override
+        public void handle(S2CTakeItemMessage packet, Player player, Object gameInstance) {
+            AddEntriesHandler.addPickUpEntry(packet.itemId, packet.amount);
+        }
+    }
 }